server/requirements.txt
fastapi[standard]==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pandas
databases[sqlite]==0.9.0
numpy==2.1.1
scikit-learn==1.5.1
regex==2024.7.24
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-multipart==0.0.9